requirements.txt
apache-airflow==2.5.0
zipfile36==0.1.3
pandas==1.3.5
python-dotenv==0.21.1
kaggle==1.5.12